Improve MCP server: bug fixes, refactoring, and comprehensive tests
- Fix entity resource URI parsing to validate type segment matches entity
- Add views.yaml to file watcher for automatic reload on changes
- Fix silent delete errors: log warnings instead of suppressing errors
- Add pagination (limit/offset) to list_relations tool
- Fix inconsistent validation error messages: extract shared validateEntity helper
- Split monolithic tools.go (1400+ lines) into focused handler files
- Extract reusable filterEntities and applyPagination[T] helpers
- Add comprehensive handler tests covering entity CRUD, relations, tracing,
  analysis, schema, resources, and helper functions (~50 test cases)
- Add watcher tests for watchDirRecursive and lifecycle
